Last try:

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -15,7 +15,7 @@ size_t print_listint(const listint_t *h)
 
 	while (h != NULL)
 	{
-		printf("%d\n", h->n);
+		printf("%d\n", (int)h->n);
 		h = h->next;
 		count++;
 	}
